refactor: modularize assistant and screenshot
